pyproject.toml
[tool.poetry]
name = "pay-api"
version = "0.1.0"
license = "MIT"
description = "API for receivables record"
authors = ["Jean Carlos Danese <jean@danese.com.br>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pydantic = "^1.9.1"
fastapi-mail = "^1.0.9"
user-agents = "^2.2.0"
email-validator = "^1.2.1"
python-multipart = "^0.0.5"
validate-docbr = "^1.10.0"
tldextract = "^3.3.0"
aiofiles = "^0.8.0"
python-dateutil = "^2.8.2"
Pillow = "^9.1.1"
phonenumbers = "^8.12.50"
psycopg2-binary = "^2.9.3"
databases = {extras = ["postgresql"], version = "^0.6.0"}
SQLAlchemy = "^1.4.37"
alembic = "^1.8.0"
PyPika = "^0.48.9"
aioredis = {extras = ["hiredis"], version = "^2.0.1"}
PyJWT = "^2.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
boto3 = "^1.24.13"
gunicorn = "^20.1.0"
qrcode = "^7.3.1"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
debugpy = "^1.6.0"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
asyncio = "^3.4.3"
httpx = "^0.23.0"
asgi-lifespan = "^1.0.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"
